+++ config.go
@@ -6,7 +6,7 @@
 
 package goat
 
-const version = "v0.7.1"
+const version = "v0.7.2"
 
 var (
 	uaString      = "goat " + version